# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api docs for lizo_alone",
        "title": "lizo_alone",
        "contact": {},
        "license": {
            "name": "all right"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/manage/article/delete": {
            "post": {
                "description": "删除文章",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除文章",
                "operationId": "删除文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章id",
                        "name": "article_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/com_model.CommonReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/manage/article/get_by_id": {
            "get": {
                "description": "获取一个文章信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "通过id获取文章",
                "operationId": "获取一个文章信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/com_model.CommonReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/manage/article/query": {
            "get": {
                "description": "查询文章列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "查询文章列表",
                "operationId": "查询文章列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索内容",
                        "name": "search_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页索引",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/com_model.CommonReturn"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/com_model.PageData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Article"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/manage/article/update": {
            "post": {
                "description": "更新文章内容",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "更新文章内容",
                "operationId": "更新文章内容",
                "parameters": [
                    {
                        "description": "文章内容",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateArticleModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/com_model.CommonReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/manage/setting/get_setting_all": {
            "post": {
                "description": "获取所有配置项",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置项"
                ],
                "summary": "获取所有配置项",
                "operationId": "获取所有配置项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索值",
                        "name": "searchKey",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/com_model.CommonReturn"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Config"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/manage/setting/update_config": {
            "post": {
                "description": "更新配置项",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置项"
                ],
                "summary": "更新配置项",
                "operationId": "更新配置项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "配置值",
                        "name": "value",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/com_model.CommonReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "com_model.CommonReturn": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "description": "数据值",
                    "type": "object"
                },
                "detail": {
                    "description": "可由运行模式控制是否显示",
                    "type": "string",
                    "example": "错误的详细信息，用于排查错误"
                },
                "message": {
                    "type": "string",
                    "example": "错误信息"
                }
            }
        },
...
        "com_model.PageData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "列表数据",
                    "type": "object"
                },
                "page_info": {
                    "description": "分页信息",
                    "type": "object",
                    "$ref": "#/definitions/com_model.PageInfo"
                }
            }
        },
        "com_model.PageInfo": {
            "type": "object",
            "properties": {
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "description": "总数据条数",
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "model.Article": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "description": "描述+关键值",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "pv": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Config": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateArticleModel": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}